Gofer: Remove update of categories
Gofer>>updateCategories seems to do nothing in the end. 
It create a new organization definition but does not use it. It just call the post load on it with a nil parameter but the post load does nothing of this class. 

So let's see if we can remove it and have one less place dealing with categories.

I also removed some related Metacello code dealing with this method