When opening a Debugger, spawn a new UI process if the existing one may be blocked
Specifically, if it is waiting on a Semaphore that is not part of a Delay, and thus may be dependent on the operation that is about to be debugged (e.g. due to waiting for an HTTP response from an in-image web server, a background operation in a threadpool, etc).

Addresses the unexpected/problematic part of the behavior mentioned in #10136