Kernel: Speedup #isObsolete
To know if a class is obsolete we are currently saving the info in class properties. But this is slow when we ask the information a lot, which happens in RPackage.

Here is an alternative where I store the obsolete classes in a weak identity set which is much faster. 

I also shipped a speed up of RPackage>>#includesClass: that should make class compilation faster.