Refactor: remove unnecessary superclass
`RBComposeParametrizedTest` was used when transformation and refactoring
had different API. More specificaly, it was used when one of them was
dealing with multiple instances and other with only single instance.
There was only one scenario where this class was used. That scenario is
remove method. Refactoring was able to remove multiple methods,
and transformation was removing only a single method.

After refactoring the `RBRemoveMethodRefactoring` the transformation
and refactoring now same the same API and this testing approach doesn't
make sense anymore.