Fix: add custom applicability precondition for remove hierarcy
This exists because remove hierarchy creates refactorings only
if method is defined in the class. Problem here is that if you
give this refactoring only methods that don't exist it won't
produce any refactorings and will just finish without changes.

This was spotted by a test that requires explicit failure in case
when method doesn't exist. We should figure a better way
of doing this in the future, since here we could reuse
applicability checks of refactorings that compose
remove method hierarchy.