RPackage: Update class category in class builder
When classifying a class in the class installer, the category is not updated. This is done later by RPAckage while listening to the announcements. But this creates a ping pong: Installer classify with the category -> RPackage listen and try to update the tag -> The tag notices a missmatch of categories -> the tag update the category -> the category is updated and an announcement rasied -> rpackage listen to it to update the tag -> since the category matches, the categorixation happens

I would like to simplify this by setting the category between the categorization and the raise of the announcement. 

Also, this will ensure the category of the class and the system organizer are in sync during the class installation which can simplify the state of the system

As a bonus I added an escape in Class>>category to manage in a better way the unclassified classes. Since a recent change I made, classes with unclassified category cannot be in a category of the system organizer. So we do not need to scan it which should speed things up in that case