RPackage: simplify package renaming
Here is a simplification of the package renaming.

Before we were checking every tag that was starting with the package name to update the name. But a tag should not be prefixed by the package name. 
The only case we have to manage is the root tag that is currently the name of the package. 

I introduced two changes with this cleaning:
- The renaming of the root tag is announced while it was silent in the past
- If we set the category to a class, but the class already has this category, we do not proceed (which will cut some update code before it happens since there is nothing to update)