Remove category ivar from MCClass(Trait)Definition
I started to push the usage of #packageName in MCClass(Trait)Definition and now most users are using #packageName and #tagName instead. The last users have a PR waiting to fix them. It is not time to remove the category instance variable. For now I'm keeping the `category:` method for backward compatibility purpose (I'll deprecate it later). And we still have too many users of #category to remove this one yet.