Fix: composite refactoring late initialization of instnace variables
When having composite refactoring that has one refactoring depending on another one
it is useful and needed to delay instantiation of instnace variables until the the execution
time. For example here we first create new temp class and later we rename that temp class.
If we would initialize the rename upfront the class is `nil` since it hasn't yet been created.
So a small hook is created that enables us to prepare refactoring for execution.