Kernel: Always recompile silently methods
While updating a class, it migrates all the methods from the old class to the new one. In that case, the methods are added silently. Well... Usually. 

In the case of a class that has a trait, the methods on the class side are not added silently but with MethodAdded announcements. 

This change removes the annoucement to always do things silently while recompiling