Unify class added and recategorized actions
This change has two main goals.

The first is to factorize the code managing announcements of class addition and recategorization. Currently it was dealt with in two places: Class>>category: and ShiftClassInstaller>>recategorize:to:. Now this is only in #category: and the second call the first one simplifying the code.

The second goal is to unify what is happening in RPackage when a class gets added or recategorized because the code was different until now but there is no reason to act differently.

A future goal is to not do the package update through announcements but this currently breaks the bootstrap