ShClassInstallerTest>>#testInstallInSpecificEnvironment should clean itself
ShClassInstallerTest>>#testInstallInSpecificEnvironment is generating a class in another environment but the tearDown only take care of the global environment. 
This change also cleans the other environment. 

Maybe we should add a #finilize method on RPAckageOrganizer to remove from the system the classes?

I think this might be the root of some failings in the STON tests