Monticello: do not rely on MCClassTraitDefinition
The infos in MCClassTraitDefinition should go in MCTraitDefinition that already have everything needed about the class trait such as slots. 

Having this duplicated info creates problem because it removes what is already loaded by MCTraitDefinition such as the class slots. 

I propose to just rely on the MCTraitDefinition. This requires also a change in Tonel that I am doing in parallel