Backport: Clean installation of MCClassTraitDefinition
This is a fix for issue #14749 in Pharo 11. 

The fix is different than in Pharo 12 because in Pharo 12 Iâ¯did cleaner version of the clean by removing the usage of MCClassTraitDefinition alltogether since all infos are already on MCTraitDefinition.

The workaround here is to not recompile entirelly the class trait if the base trait is already present in the image. Instead it will just update the trait composition, thus it will not remove the instance variables.