RPackage: Implement UndefinedPackage
I did a PR to introduce an UndefinedPackage and UndefinedPackageTag but this seems to be a lot of changes at once and breaks a lot of thing.

Here is a smaller set of changes that only introduces UndefinedPackage and do not change anything to the behavior if I did things right.

I also renamed #isDefault into #isUndefined.

Depends on #14784
Subset of #14778