RPackage: Update meaning of #removePackage:
In the past we had #unregisterPackage: that was unregistering a package from the package organizer. Often the vocabulary about registration was missused and I removed all those method to use "add/remove" instead. But in the case of this particular method I think I mess up because it is really unregistering in the sens that it does not remove the content of the package. 

To replace it I introduced #removePackage:. In this commit I add back #unregisterPackage: to do the unregistration but as a private method that should not be used if possible. And I updated #removePackage: to remove also the content of the package to remove. 

I updated all the senders in the image to use the right one. This will leave less orphan classes in the system.