Fix announcements during class creation
If we create a class with traits in the system, the class builder is installing some basic methods in the class. This operation should be silent but in fact we are raising some announcements currently. 

This fixes the problem and adds a tests to ensure the behavior is not reverted.