RPackage: remove private import method
Remove RPackage>>importClass:inTag: because this private method is called only in one place in RPackage and the fact that it is here made some people want to use it. Instead people should use #addClass: to add a class to a package without caring about the tag or #moveClass:toTag: if they care about the tag. Like this, we are removing the class from the old package/tag and announcing the repackaging