RPackage: ClassRepackaged announcement should know the package tags
ClassRepackaged is fired when a class change of package OR when the class change of tag. But this announcement is configured only with the package. Because of this, some classes are relying on ClassRecategorized instead of ClassRepackaged. (I want to remove ClassRecategorized)

This change adds tags info in ClassRepackaged so that we can really remove ClassRecategorized