RPackage: #moveClass:toTag: can be simpler in the package do not change
Currently #moveClass:toTag: is acting as if the package of the class always change. This updates the behavior so that we do a full update only if the package change. In case only the tag changed, we can just remove from the old tag and add to the new one.