Clean test on repackaging
SlotAnnouncementTest contains a test that has nothing to do with slots about recategorization of methods.

I removed it and instead I implemented a new test in ShClassInstallerAnnouncementsTest to test teh repackaging. It is in a better place than the previous test and rely on repackaging and not recategorization that I am trying to remove.
I also removed some things in the tearDown methods that should not be needed anymore since a recent fix I did in RPackage