Update DoItChunk to manage new package API
DoItChunk is used a lot to import code from Monticello and often this is to import a package or a class.
The API used to import package had been deprecated and should be removed, but this might break all the old code export of Monticello. 

This change updates the code importer to check if we are in that case, and if we are, it uses the new API instead of evaluating the code.