Add tests on Fluid class parser and fix bugs
This changes aim to add tests on the creation of classes from the FuildClassDefinitionParser.  I ported the tests from ASTClassBuilderTest and added a missing one on nil management.

I also fixed a few bugs:
- Shared pools parsing was broken
- nil as a receiver was not working
- Trait as a receiver was not working

For now I keep the tests on the old class parser in ASTOldClassBuilderTest because this is still used by Calypso. We can remove them once we remove the `[v] Fluid`  checkbox from Calypso.

On other subject, I migrated UndefinedClassTest to use the fluid class definition.

Fixes #14979