Remove BaselineOfIDE obsolete management
It seems we have no obsolete anymore and if one is added a release test should catch it. So I propose to remove this. Also I propose to remove setting a new author that seems useless. 

But doing this change I see something really bad... We load Iceberg and extra tools from the post load and not from the baseline! This is bad :(