Improve disabling of Epicea in tests
In Pharo 12 I disabled Epicea in tests except for Epicea tests. But I am not satisfied with how I did it.

I did it by preventing to add events to the EpMonitor when the execution environment was the test environment. But that has the problem that Epicea is still listening to the events and generating its changes. This can be slow and in some edge cases creates bugs (for example when you test some things on another environment than the default environment). 

I propose to cut Epicea right from the start: when we receive the announcement.