Improve new chunk importer
During my update of the chunk importer I did an approximation for the packages and tags considering that what was after the last dash was the tag. But it is possible to keep the old behavior by checking the packages in the package organizer.

This should help to get rid of the categories because the bootstrap is still relying on the chunk importer :(