Simplify bootstraping of Metacello
Metacello is currently part of Pharo and comes in its full version in Pharo images which was not always the case in the past. In case we did not had the full Metacello in the past, there was a bootstraping of Metacello happening to load what we wanted. 

This bootstraping is not needed anymore and I propose here to remove all the code related to this to make things simpler.

I aslo simplified some code in Metacello-Base using #ifEmpty: for example instead of `isEmpty ifTrue:` etc

As a next step I would like to merge Metacello-Base in Metacello-Core since the code depend on it and they are always loaded together