Fix bug marking too many packages as dirty
Too many packages end up bee marked as dirty while loading code in Iceberg. I traced the root of one of the problem happening. This problem comes from MethodAdded announcement that is pointing the wrong package as the package contaning the method.

I propose to not save in a variable the package of a method in the method annoucements (except for the MethodRemoved for now) but to jutst ask the method. 

I added a regression test about this behavior. I think this problem will be solved in an even cleaner way in the future when the MethodAdded will not be managed through announcements but directly in the code of the methods.