RPackage: Simplify protocol management
RPackage currently listens to all announcements related to protocols and when they receive one they ensure we have a package for every extension protocol of the class. 

This is doing too much work. First we are not doing anything when a protocol gets removed. And when a protocol gets added, we don't need to do things for every protocols of the class, just for the one that got added.

This cfange simplifies what RPackage is doing on protocol management