Monticello: Make MethodAddition closer to system behavior
Monticello reimplement its own method compilation mecanism and I don't like this because it means that we have duplication and this can (and already has in the past) introduce bugs. 

I tried multiple things in the past to remove this duplication but it is not so easy. Here I'm trying the little step of making the announcements the same than the ones of the system.