Improve MCSystemCategoryParser
In the past months I did some work to remove the categories from the system. During that cleaning I updated MCSystemCategoryParser but the new version does not act as the old one. Now that I'm more familiar with the system I can update it again to get back some of the old behaviors