Introduce and use CategoryConverter
With the removal of the categories we need to handle the code that was using categories in the past. I already handled it in multiple places but this was duplicating the behavior and I found even more places.

In the end I'm proposing to introduce a CetagoryConverter class to centralize this behavior and make all other places rely on it. I also updated MCStReader and MCTraitParser to handle better the categories