Move traceCr & co to non interactive transcript
We have Trancript-Core (which is a bad name IMO. It should be Transcript-Interactive) and Transcript-NonInteractive. The second one is loaded in the bootstrap while the second one is loaded with Morphic. 

The first one is adding #traceCr and co to Object but that makes it unavailable in the bootstrap. I propose to move them to non interactive transcript so that we can log things on the stdout ealier with it.