Improve the management of TTranscript
We have in the system Transcript-Core-Traits and it has 2 problems IMO.

The first one is that TTranscript (the only class inside) isn't used.

The second is that it is loaded in the first steps of the bootstrap.

Since Transcript-Core is loaded at a point where we have the traits in the image I propose to move TTranscript to this package (no need to have a package for only one class..) and we should make ThreadSafeTranscript use it