RPackage: Clean package management
First I removed the usage of announcements to update RPackage.

I simplified some code in Calypso.

And I removed a usage of #ensurePackageOfExtensionProtocol: to use #packageForProtocol: because the package should always be created before and I want the code to fail if it is not the case so that we find the problem