Remove #acceptsLoggingOfCompilation
This method is a hook that is never used in the image. If someone wants to not log, Opal offers an API for that and if a class want to avoid logging, it can overrid #compiler on the class side to set the logging to false. It will have the same effect