Inline date printer
BasicDatePrinter and ExtendedDatePrinter introduce problems of depenedencies. BasicDatePrinter is never used because its #default method is returning ExtendedDatePrinter. On top of this, we have a dependency problem because ExtendedDatePrinter is in a package that is loaded later in the bootstrap. 

I propose to just inline things to make the code simpler. I deprecated those two classes