Improve logs while running tests
The goal of the change is to improve logs while running tests.

It prints when it starts, when it ends and how long it took.

Also it now uses the Transcript instead of Stdout to be sure the logs are in the right order (because transcript is buffered). Also I flush often because if a test crashes the image while the buffer of the transcript is not flushed. we cannot know when the problem occured.