Calypso: Simplify ClyExtendingPackagesQuery adn clean package API
This change mainly simplifies the code of ClyExtendingPackagesQuery to make it easier to understand and faster (At least it will be faster when #15212 will be integrated).

I also deprecated #isDefinedInPackage: because it the bahavuior was unclear and it is easy to do without this method polluting the API (and it is not used).

Thie behavior was unclear because in some part of the systemit returned true if the entity was in the package as parameter while in other part it returned true if the entity was in the package AND was not an extension (extensions could never return true)