Clean traits tests
Traits tests seems to leave the system in a bad state now that the categories are not used to install a class anymore. This is probably due to that fact that the class builder was not using the right package/tag before.

So here is a cleaning. On top of fixing the problem I also simplified the cleaning code and the API to create classes and traits. I used the vocabulary that we are using in ShiftClassBuilder now. I also renamed a class to remove the "category" vocabulary that is been removed from the image slowly.