Calypso: Add Undefined project
In case some packages of the image are in no project, I'm adding an undefined project. 

This will allow me to implement a new feature more easily: Select package in projects. Currently in the project view if you select a class of a package that is not part of the opened project (with the spotter for example), it will go back to the packages view. I want to select the first instance of the package in the view but for that I need to be sure all packages are present in the view.

This is a first step