Fix trait method recompilation and clean
The main goal of this PR is to fix the fact that recompiling a method from a trait does not update the method in the users to install the recompiled method. 

I also found some things that seems inconsistant and tried to clean them. All the traits tests are passing so I guess this is good?

Changes:
- CompiledMethod>>#recompile now ask to the origin to recompile the method and not the method class
- Recompiling on TraitedClass and TraitedMetaclass ask the TraitChange class to add the updated method to the users method dictionary
- Unify the recompilation code between TraitedMetaclass and TraitedClass. Here I am not sure about what I did! Maybe we should keep the code on the #traitSource property but add it also to TraitedClass?
- Remove TraitedClass class>>#addAndClassifySelector:withMethod:inProtocol: because I have the impression that the version on TraitedMetaclass is already managing this

I also added some tests

Fixes #15202

Co-authored-by: MarcusDenker <marcus.denker@inria.fr>