Introduce a specific collection for undeclareds
The goal of this change is to introduce a special collection for Undeclareds. This will allow us to add some constraints to what an undeclared is. For example we can ensure that undeclared contains only UndeclaredVariables and not an association.

This also fix a test that was adding association in the undeclareds