Packages: Remove #definedSelectors cache
#definedSelectors cache is a dictionary of the classes defined in a package and the method selectors defined in those classes. 

This is used because to know the defined selectors of a class you need to reject the methods from traits and the extension methods and #isExtension was way too slow. 

Recently I did a huge speed up of #isExtension so we are now able to remove this cache and simplify the code! I also did a speed if of #includesClass: 

This PR just deal with the removal of the cache but we will be able to also remove more mecanism after this change. I'm just trying to do it step by step to not break the system. This will also avoid any synchronization missmatch between RPackage and Classes