Opal: Make plugins no-op to not crash during the loading
It is possible that installing a package with an opal plugin and a class using it in the same package crashes because all methods of the plugin are not loaded. 

This change make Opal pluggin superclass a no-op to avoid crashes on simple cases and add a warning in the class comment to warn about this.