Package: Move extension methods of RPackage
This week I'll integrate for real RPackage in the MM. I wanted to do it in one step but I'm noticing that it is way too big. Impossible to review and probably a lot of rooms for stupid mistakes. So I'll split the change into smaller ones and here is the first step. 

Methods such is ClassDescription>>#package and company should not be extensions. 
I also cleaned some code to make it more readable or remove accesses to globals (like in SmalltalkImage>>#package)