Traits: Fix bug in dictionary rebuild
Ok so this one was not an easy one and took me quite a while to hunt.

When we remove a trait from the system, it removes itself from its users. When this happens, the method dictionary of the behavior is rebuilt and we had a bug there!
The bug was that TaCompositeElement>>#selector was returning the selectors of the methods of the element. But since we were in the middle of the method dictionary rebuild, we did not remove the copied methods from the removed trait yet! (Or maybe we got lucky on the rebuild order and it worked! This is the origin of the bug randomness). 

I updated the method to collect the selectors of the trait composition instead. Since the trait composition is right, the list of selectors will be updated even before we rebuild the method dictionary.

I also fixed another bug were diamonds in trait composition could return multiple times the selectors. 

Fixes #15282

Side note: Why did it start to fail now? Because before we were removing the classes in the reverse order of their compilation by a piece of code that was done by hand. Now we delegate this to the system by removing the packages containing the generated code. This means that the order is not the creation order. And depending on this, it could fail or not. But with this fix, we fix it for all trait removals and we do bypass this bug by taking care of the order ourself in the tests