Auto completion should propose globals
Since a few years it frustrate me that the auto completion does not propose the globals names. It's bad to use globals but sometimes you don't have the choice, and it can make our like much better while debugging. For example to use the transcript.

Here is a fix so that the CoGlobalVariableFetcher fetch global variables and not just classes