Fix Iceberg detection of extension methods removal
Since a while (and checking the code it is probably since a loooong time), if you remove a class with extensions, the extension packages are not marked as dirty in Iceberg.

The origin was that ClassRemoved>>#packagesAffected did not returned the extending packages. I'm fixing this here and adding a regression test. 

I also simplified the code of Monticello. The code to all classes and methods announcement can be factorized into one method now that the announcements got fixed