Fix bug in recompilation of trait methods
Methods copied from traits have a #traitSource property. This property can be lost on recompilation. 

This change adds tests and fixes this. For now the properties to persist are hard coded. In a next iteration I'll make this configurable (in fact I already have a PR to try and do that but it is not working yet)