Ring: Fix RGPackage>>#asMCSnapshot
I am working on using the ring version of the image to run Iceberg tests and avoid to depend on an unmaintained external project. Here is a fix to be able to use the in image version of Ring.

RGPackage>>#asMCSnapshot was not creating a MCOrganizationDefinition which is necessary to load a package