Add los on PharoBootstrapInitialization
PharoBootstrapInitialization is doing the initial initialization of the Pharo image after Espell finished to initalize the minimal image. It happened to mu multiple times that this step failed on some of my PR and it is really hard to debug. Here I am adding some logs to be able to pinpoint more easily where the build is failing and help to debug.

I'm also trying to group the initializations that makes sense to be together