Bootstrap: Update localisation of SourceFileArray
I am trying to remove the dependencies of the core packages of Pharo (the ones dealt with via Espell) over the SessionManager because we could load it via Hermes.

Here is a cleaning moving the initialization of source file array in the PharoBootstrapInitialization class. Also this is now done in the Pharo code and not the bootstrap bash scripts.