Refactor: DeprecateClassTransformation only adds `isDeprecated` to the class side
Previously, this transformation overrided `new` and `systemIcon` methods that
are not needed to override.
Also, it was incorrect since it was creating `deprecated` method which doesn't
not make the class deprecated.
Finally, we've removed comment creation since new comment overrides exisitng
one, and it's currently very hard to preserve the old comment and prepend it with
"Deprecated". But the system strikes through the class, so we have visual
representation, which should be enough.