Introduce Kernel-CodeModel package
This change splits Kernel package to introduce a Kernel-CodeModel package. 

Kernel is a really big package, the goal is to improve the organisation of Pharo to make it easier to find what we are looking for. Packages, Classes, Methods, Blocks, Protocols and related classes are entities of the same concept and could be packaged together