Traits: Fix update of source pointers
It seems that traits were setting the source pointers of the copied methods in system dictionary while adding the method to the class. 

A problem is that the logging of the source code now happens after we install the methods in the system because some info were missing if we logged this before. This result in the fact that currently we have copied methods missing their source pointer. 

I propose an alternative here by updating the source pointer of the copied methods in the users upon the update of the source pointer in the original method instead of doing it while adding the method to the user.