Traits: Fix wrong source pointer
This is an alternative to https://github.com/pharo-project/pharo/pull/15732

The bug was introduced while moving the logging of methods after the installation. This change reverts it. This was done because sometime we cannot know the actual protocol of a method before classifying it. But this protocol is not used currently so it is better to have the traits working than to fix a problem that currently does not harm us much