Return an OrderedCollection from SortedCollection>>join:
I would argue that `#(3) asSortedCollection join: #(1 2)` should answer `OrderedCollection(1 3 2)` not `SortedCollection(1 2 3)`—that the interleaved order of the result is an important part of the intent of `join:`, and this takes precedence over the usual preference to return the same class as the receiver. It's also worth noting that, with the old implementation, a SortedCollection with a custom sortBlock didn't copy that sortBlock to the output of `join:`, which seems like the worst of both worlds—sorted, but _differently_ from the receiver? Or it's likely to error if the elements don't understand `#<` and the sortBlock was sorting on some property.