Position temp definition before first occurrence of extracted expression, and always in outer scope (fixes #14611)
If the first occurrence of the expression to be extracted is in a child scope (a block) of the occurrence that was actually *selected*, we need to walk up the tree and put the initialization before the whole statement containing that child scope.

Add RBProgramNode>>statementNodeIn: to answer the statement *at a particular level of nesting/in a particular scope* that contains a node.