Background: in ST80, you have to check isBits first before you can check for bytes/words to get a meaningfull result. (isWords was defined as "isBytes not" (see https://github.com/pharo-project/pharo/issues/15668 for an open issue discussing that).
Another PR fixed that and just delegate to the layout, but there the isBytetes and friend where only defined for BitLayout (to force a check for isBits fist to mimick the implentation on Behavior).

This PR Moves the isBytes/isWords methods up in the Layout Hiearchy, now it is ok to call the without checking for isBits first.

fixes #15941