Speedup Float and Integer comparison
I had a strange case while optimizing a clustering algorithme were an implementation was really slow using `x < y` in the code. And it became almost instantaneous using `y >= x`. This is because x is usually a float and y an integer. 

In the first case the comparison primitive fails and the fallback casts the float as fraction. In the second case, the primitive works directly.

I propose to speed up #>, #<, #>= and #<= by inverting the order of the comparison and making the integer the receiver instead of the float. 

This causes a little slow down on #= and #~= in case the fraction part is not 0.0 but this seems relatively small compared to what we gain

```st
[
value := 1.00000001.

10000000 timesRepeat: [
	value < 1.
	value := value + 1 ] ] timeToRun

 "Before: 0:00:00:02.38"
 "After: 0:00:00:00.201"
```